fix en migraciones

+++ database/migrations/2024_07_29_162451_create_conta_grupo_cuenta_resultado_rpt_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('cuenta_id')->nullable();
             $table->string('codigo', 200)->nullable();
             $table->unsignedBigInteger('sub_grupo_id')->nullable();
-            $table->unsignedInteger('grupo_id')->nullable();
+            $table->unsignedBigInteger('grupo_id')->nullable();
             $table->foreign('sub_grupo_id')->references('id')->on('conta_grupo_sub_resultado_rpt')->onUpdate('cascade');
             $table->foreign('grupo_id')->references('id')->on('conta_grupo_resultado_rpt')->onUpdate('cascade');
             $table->timestamps();
